requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
supabase==2.6.0
openai==1.35.10
APScheduler==3.10.4
pytz==2024.1
uvloop==0.19.0
orjson==3.10.7 